% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdec.R
\name{make_fs_qdec}
\alias{make_fs_qdec}
\title{Create a Freesurfer qdec}
\usage{
make_fs_qdec(data, formula, path = NULL, keep = FALSE)
}
\arguments{
\item{data}{data.frame, list or environment
containing the data. Neither matrix nor an array
will be accepted.}

\item{formula}{a model \code{\link{formula}} or
\code{\link{terms}} object. For simple qdec file,
with binary columns for all levels of a factor,
make sure to include a formula with \code{-1} to
remove the intercept.}

\item{path}{an option file path to write qdec to csv}

\item{keep}{logical or vector of column names,
to keep the original data in the output.
Default is \code{FALSE}.}
}
\value{
data.frame with model matrix and
scaled continuous variables.
}
\description{
Create a Freesurfer qdec from a model formula,
utilising R's robust model formula syntax. The function
also scales continuous variables, and can keep the original
data in the output.
}
\details{
The \code{\link{formula}} should in all likelihood also
include the \code{-1} to remove the intercept, as this
will provide a matrix where all levels of factor variables
have their own binary column. This is necessary to follow
the steps from the Freesurfer documentation.

It is highly recommended to have the base-id's for Freesurfer
in their own column, and request to have this remain
in the qdec by using the id-column's name in the
\code{keep} argument.
}
\examples{
cars <- mtcars
cars$cyl <- as.factor(cars$cyl)
cars$gear <- as.factor(cars$gear)

make_fs_qdec(cars, mpg ~ cyl + hp)

# Remove the intercept, necessary to follow
# steps from Freesurfer docs
make_fs_qdec(cars, mpg ~ -1 + cyl + hp)
make_fs_qdec(cars, mpg ~ -1 + cyl + hp + gear)

# Keep the original data also in the output
make_fs_qdec(cars, mpg ~ -1 + cyl + hp, keep = TRUE)

# Keep the original data of specific columns
# Use a character vector
make_fs_qdec(cars, mpg ~ -1 + cyl + hp, keep = c("mpg", "gear"))

}
