% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdec.R
\name{plot.qdec}
\alias{plot.qdec}
\title{Plot qdec matrix}
\usage{
\method{plot}{qdec}(x, col = hcl.colors(12, "viridis"), ...)
}
\arguments{
\item{x}{a qdec object}

\item{col}{a vector of colors to be used in the heatmap.}

\item{...}{arguments to be passed to \code{\link{heatmap}}.
\code{scale}, \code{Rowv}, \code{Colv} already have
custom values that may not be overwritten for the sake
of a better visualisation.}
}
\value{
a heatmap of the qdec matrix
}
\description{
Visualise a Freesurfer qdec matrix as returned
byt the \code{\link{make_fs_qdec}} function.
}
\examples{
cars <- mtcars
cars$cyl <- as.factor(cars$cyl)

qdec <- make_fs_qdec(cars, mpg ~ -1 + cyl + hp)
plot(qdec)
}
