% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdec.R
\name{qdec}
\alias{qdec}
\title{Freesurfer qdec constructor}
\usage{
qdec(data, formula)
}
\arguments{
\item{data}{input data.frame}

\item{formula}{a model \code{\link{formula}} or
\code{\link{terms}} object. For simple qdec file,
with binary columns for all levels of a factor,
make sure to include a formula with \code{-1} to
remove the intercept.}
}
\value{
qdec with model matrix
}
\description{
Creates a qdec matrix from a model formula.
}
\examples{
cars <- mtcars
cars$cyl <- as.factor(cars$cyl)

qdec <- qdec(cars, mpg ~ cyl + hp)
}
