% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_num_data}
\alias{scale_num_data}
\title{Scale numeric variables in a data frame}
\usage{
scale_num_data(data, vars)
}
\arguments{
\item{data}{data.frame}

\item{vars}{variables selected}
}
\description{
Scale numeric variables in a data frame
}
