% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_vec}
\alias{scale_vec}
\title{Scale a numeric vector}
\usage{
scale_vec(x, ...)
}
\arguments{
\item{x}{numeric vector to scale}

\item{...}{additional arguments to be passed to \code{\link{scale}}}
}
\value{
scaled vector
}
\description{
Scale a numeric vector
}
\examples{
scale_vec(1:20)
}
